Move range proofs to own file
